+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `NavisworksContentsBuilder` to build Navisworks `AppBundle`
 - Add `IssNavisworksBuilder` with `EnableUserAppData` to install in current `UserAppData`.
 - Update `IssAppBundleBuilder` to use `PrivilegesRequired.Lowest` when install with `UserAppData`.
+- Add `FileEditorExtension` extension to edit license files to replace keys like `$AppName$` and `$Year$`
 ### Tests
 - Add `AutoCADExtension` tests with `AutoCADAddin` files for version `2019` to `2026`.
 - Add `NavisworksExtension` tests with `NavisworksAddin` files for version `2021` to `2026`.